pyproject.toml
[project]
name = "ai-search-with-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.13.0",
    "azure-identity>=1.25.1",
    "azure-identity-broker>=1.3.0",
    "azure-search-documents>=11.6.0",
    "fastmcp>=2.12.4",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "ipykernel>=7.0.1",
    "pandas>=2.3.3",
]
